# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Infinity Knowledge Control Plane API
  version: 1.0.0
  description: |
    Machine-readable contract for the implemented `/api/v1` HTTP surface.
    Every operation requires an OIDC bearer token. Operations described as
    administrator-only additionally require the `knowledge-admin` realm role.
    Tenant and principal identity are derived exclusively from the validated token.
servers:
  - url: http://localhost:8080
    description: Local control plane
security:
  - bearerAuth: []
...
paths:
  /api/v1/spaces:
    post:
      tags: [Spaces]
      operationId: createSpace
      summary: Idempotently create a tenant knowledge space
      description: Requires `knowledge-admin`; the creator receives an administrator ACL grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpaceRequest'
      responses:
        '204':
          description: Space exists and has been provisioned
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/admin/overview:
    get:
      tags: [Management]
      operationId: getAdminOverview
      summary: Get tenant management counters
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Overview counters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/spaces:
    get:
      tags: [Management]
      operationId: listAdminSpaces
      summary: List tenant spaces for administration
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Tenant spaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementSpace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/documents:
    get:
      tags: [Management]
      operationId: listAdminDocuments
      summary: Page tenant documents for administration
      description: Requires `knowledge-admin`; deleted documents are omitted.
      parameters:
        - name: spaceId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Document page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/documents/{documentId}/chunks:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Management]
      operationId: listAdminDocumentChunks
      summary: List active-revision chunks for one document
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Chunks ordered by ordinal
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementChunk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/documents/{documentId}/revisions:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Management]
      operationId: listAdminDocumentRevisions
      summary: List immutable revisions for one document
      description: Requires `knowledge-admin`; the active revision is identified by the `active` field.
      responses:
        '200':
          description: Revisions ordered by revision history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementRevision'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/documents/{documentId}/revisions/{revisionId}/chunks:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - $ref: '#/components/parameters/RevisionId'
    get:
      tags: [Management]
      operationId: listAdminDocumentRevisionChunks
      summary: List chunks for one immutable document revision
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Revision chunks ordered by ordinal
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementChunk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/connectors:
    get:
      tags: [Connectors, Management]
      operationId: listAdminConnectors
      summary: List tenant connector configurations and last-run summaries
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Connectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManagementConnector'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/traces:
    get:
      tags: [Traces]
      operationId: listTraces
      summary: List recent tenant retrieval traces
      description: Requires `knowledge-admin`.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Retrieval traces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/traces/{traceId}:
    parameters:
      - $ref: '#/components/parameters/TraceId'
    get:
      tags: [Traces]
      operationId: getTrace
      summary: Get one tenant retrieval trace
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Retrieval trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trace'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/admin/wiki/pages:
    get:
      tags: [Wiki]
      operationId: listWikiPages
      summary: List governed Wiki pages
      description: Requires `knowledge-admin`.
      parameters:
        - name: spaceId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WikiPageStatus'
      responses:
        '200':
          description: Wiki page summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WikiPageSummary'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Wiki]
      operationId: compileWikiPage
      summary: Compile selected active chunks into a Wiki draft
      description: Requires `knowledge-admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompileWikiPageRequest'
      responses:
        '201':
          description: Compiled Wiki page detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiPageDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/WikiConflict'
  /api/v1/admin/wiki/pages/{pageId}:
    parameters:
      - $ref: '#/components/parameters/PageId'
    get:
      tags: [Wiki]
      operationId: getWikiPage
      summary: Get Wiki page, latest revision and active revision
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Wiki page detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiPageDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/wiki/pages/{pageId}/submit-review:
    parameters:
      - $ref: '#/components/parameters/PageId'
    post:
      tags: [Wiki]
      operationId: submitWikiPageForReview
      summary: Transition a Wiki page to IN_REVIEW
      description: Requires `knowledge-admin` and optimistic version matching.
      requestBody:
        $ref: '#/components/requestBodies/WikiTransition'
      responses:
        '200':
          $ref: '#/components/responses/WikiPageDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/WikiConflict'
  /api/v1/admin/wiki/pages/{pageId}/reject:
    parameters:
      - $ref: '#/components/parameters/PageId'
    post:
      tags: [Wiki]
      operationId: rejectWikiPage
      summary: Transition a Wiki page back to DRAFT
      description: Requires `knowledge-admin` and optimistic version matching.
      requestBody:
        $ref: '#/components/requestBodies/WikiTransition'
      responses:
        '200':
          $ref: '#/components/responses/WikiPageDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/WikiConflict'
  /api/v1/admin/wiki/pages/{pageId}/publish:
    parameters:
      - $ref: '#/components/parameters/PageId'
    post:
      tags: [Wiki]
      operationId: publishWikiPage
      summary: Transition a Wiki page to PUBLISHED
      description: Requires `knowledge-admin` and optimistic version matching.
      requestBody:
        $ref: '#/components/requestBodies/WikiTransition'
      responses:
        '200':
          $ref: '#/components/responses/WikiPageDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/WikiConflict'
  /api/v1/admin/wiki/pages/{pageId}/archive:
    parameters:
      - $ref: '#/components/parameters/PageId'
    post:
      tags: [Wiki]
      operationId: archiveWikiPage
      summary: Transition a Wiki page to ARCHIVED
      description: Requires `knowledge-admin` and optimistic version matching.
      requestBody:
        $ref: '#/components/requestBodies/WikiTransition'
      responses:
        '200':
          $ref: '#/components/responses/WikiPageDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/WikiConflict'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RevisionId:
      name: revisionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TraceId:
      name: traceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageId:
      name: pageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  requestBodies:
    WikiTransition:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WikiTransitionRequest'
  responses:
    Unauthorized:
      description: Bearer token is absent or invalid
    Forbidden:
      description: The authenticated principal is not allowed to perform the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidRequest:
      description: Request validation, parsing or domain argument failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    WikiConflict:
      description: Optimistic version or Wiki lifecycle transition conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    WikiPageDetail:
      description: Current Wiki page detail
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WikiPageDetail'
  schemas:
    ApiError:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    CreateSpaceRequest:
      type: object
      required: [spaceId, name]
      properties:
        spaceId:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 256
    AccessibleSpace:
      type: object
      required: [id, name, description, status, version, documentCount, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
        version:
          type: integer
          format: int64
        documentCount:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
...
    DocumentStatus:
      type: string
      enum: [DRAFT, ACTIVE, DEPRECATED, ARCHIVED, DELETED]
...
    ManagementOverview:
      type: object
      required: [spaces, activeDocuments, chunks, connectors, pendingProjections, deadProjections, evaluationDatasets, evaluationRuns]
      properties:
        spaces: {type: integer, format: int64}
        activeDocuments: {type: integer, format: int64}
        chunks: {type: integer, format: int64}
        connectors: {type: integer, format: int64}
        pendingProjections: {type: integer, format: int64}
        deadProjections: {type: integer, format: int64}
        evaluationDatasets: {type: integer, format: int64}
        evaluationRuns: {type: integer, format: int64}
    ManagementSpace:
      allOf:
        - $ref: '#/components/schemas/AccessibleSpace'
    ManagementDocument:
      type: object
      required: [id, spaceId, title, sourceType, sourceUri, status, authority, version, activeRevisionId, chunkCount, keywordStatus, vectorStatus, graphStatus, originalFileName, sourceMediaType, sourceContentLength, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        spaceId:
          type: string
        title:
          type: string
        sourceType:
          type: string
        sourceUri:
          type: string
        status:
          $ref: '#/components/schemas/DocumentStatus'
        authority:
          type: integer
        version:
          type: integer
          format: int64
        activeRevisionId:
          type: string
          format: uuid
          nullable: true
        chunkCount:
          type: integer
          format: int64
        keywordStatus:
          type: string
        vectorStatus:
          type: string
        graphStatus:
          type: string
        originalFileName:
          type: string
          nullable: true
        sourceMediaType:
          type: string
          nullable: true
        sourceContentLength:
          type: integer
          format: int64
          nullable: true
        updatedAt:
          type: string
          format: date-time
    DocumentPage:
      type: object
      required: [items, limit, offset, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ManagementDocument'
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
          format: int64
    ManagementChunk:
      type: object
      required: [id, ordinal, sectionPath, content, contentHash]
      properties:
        id:
          type: string
          format: uuid
        ordinal:
          type: integer
        sectionPath:
          type: array
          items:
            type: string
        content:
          type: string
        contentHash:
          type: string
    ManagementRevision:
      type: object
      required: [revisionId, revisionNumber, contentHash, mediaType, language, parserVersion, createdAt, active, chunkCount]
      properties:
        revisionId:
          type: string
          format: uuid
        revisionNumber:
          type: integer
          format: int64
        contentHash:
          type: string
        mediaType:
          type: string
        language:
          type: string
        parserVersion:
          type: string
        createdAt:
          type: string
          format: date-time
        active:
          type: boolean
        chunkCount:
          type: integer
          format: int64
    ManagementConnector:
      type: object
      required: [id, spaceId, type, displayName, status, version, lastRunId, lastRunStatus, lastRunAt, updatedAt]
      properties:
        id:
          type: string
        spaceId:
          type: string
        type:
          type: string
        displayName:
          type: string
        status:
          type: string
        version:
          type: integer
          format: int64
        lastRunId:
          type: string
          format: uuid
          nullable: true
        lastRunStatus:
          type: string
          nullable: true
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
    Trace:
      type: object
      required: [id, requestId, principalId, totalDurationMs, resultCount, createdAt, steps]
      properties:
        id:
          type: string
          format: uuid
        requestId:
          type: string
          format: uuid
        principalId:
          type: string
        totalDurationMs:
          type: integer
          format: int64
        resultCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: '#/components/schemas/TraceStep'
    TraceStep:
      type: object
      required: [ordinal, name, durationMs, inputCount, outputCount, status]
      properties:
        ordinal:
          type: integer
        name:
          type: string
        durationMs:
          type: integer
          format: int64
        inputCount:
          type: integer
        outputCount:
          type: integer
        status:
          type: string
...
    WikiPageStatus:
      type: string
      enum: [DRAFT, IN_REVIEW, PUBLISHED, ARCHIVED]
    CompileWikiPageRequest:
      type: object
      required: [spaceId, slug, title, sources]
      properties:
        spaceId:
          type: string
          minLength: 1
          maxLength: 64
        slug:
          type: string
          minLength: 1
          maxLength: 256
        title:
          type: string
          minLength: 1
          maxLength: 512
        sources:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/WikiSourceSelection'
    WikiSourceSelection:
      type: object
      required: [documentId, revisionId, chunkId]
      properties:
        documentId:
          type: string
          format: uuid
        revisionId:
          type: string
          format: uuid
        chunkId:
          type: string
          format: uuid
    WikiTransitionRequest:
      type: object
      properties:
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
          default: 0
    WikiPageSummary:
      type: object
      required: [id, spaceId, slug, title, status, latestRevisionId, activeRevisionId, version, sourceCount, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        spaceId:
          type: string
        slug:
          type: string
        title:
          type: string
        status:
          $ref: '#/components/schemas/WikiPageStatus'
        latestRevisionId:
          type: string
          format: uuid
        activeRevisionId:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          format: int64
        sourceCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
    WikiPageDetail:
      type: object
      required: [page, latestRevision, activeRevision, unpublishedChanges]
      properties:
        page:
          $ref: '#/components/schemas/WikiPageSummary'
        latestRevision:
          $ref: '#/components/schemas/WikiRevision'
        activeRevision:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/WikiRevision'
        unpublishedChanges:
          type: boolean
    WikiRevision:
      type: object
      required: [id, revisionNumber, summary, markdown, sources, contentHash, compilerVersion, generatedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        revisionNumber:
          type: integer
          format: int64
        summary:
          type: string
        markdown:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/WikiSourceReference'
        contentHash:
          type: string
        compilerVersion:
          type: string
        generatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    WikiSourceReference:
      type: object
      required: [documentId, revisionId, chunkId, sectionPath, contentHash, authority]
      properties:
        documentId:
          type: string
          format: uuid
        revisionId:
          type: string
          format: uuid
        chunkId:
          type: string
          format: uuid
        sectionPath:
          type: array
          items:
            type: string
        contentHash:
          type: string
        authority:
          type: integer